Cleaning headers again and test1. Various cleanup operations. Mainly restores. Going back to the way ALL unix based programs work, that is, by invoking one executable with a simple name, e.g. charmm, sander, gmx, mdrun, resp, xleap, no more qfep5 or even worst! UPPERCASE Qfep5, as if it was a bad windows executable of sorts. This is in line with the documentation.
Take a good look at the Fortran Best Practices:

http://www.fortran90.org/src/best-practices.html

Or at the gromacs coding rules:

http://jenkins.gromacs.org/job/Documentation_Nightly_master/javadoc/dev-manual/formatting.html

In principle, following a pep8 python style in
Fortran is useful. You are not coding alone!

Still tons of clean-up work is needed for
code readability and a developers manual is
a most. I am writing in LaTeX as it is easy to
maintain and could be eventually translated directly
to html or markdown if the need arises.