HOTFIX for qcalc5
Fix for memory corruption if too many angles or torsions are broken. Previous versions could crash in sanity check for large numbers of angles and torsions because the sanity check allocated memory according to number of breaking bonds and not number of actual breaking torions and angles. Fixed by not being stupid and changing the memory allocation.