Fix for broadcasting solvent type to nodes
Major fix to make sure that the solvent type is actually send to the nodes during the MPI setup.
MPI Broadcast for solvent_type had been missing, luckily for everyone this defaults to 0 (spc-water).
Also minor fix for n-atom solvent vdW calculation to make sure vdW types are always exact and not dependent on first atom in solvent molecule.