Major update
Total rewrite of the way nonbonded interactions are computed. Now all possible interactions are first precomputed (after the initialization of the nodes on each node) and then used to both populate the nbxx lists and to have the interaction energies ready for use in the nonbond_xx functions. The routines have now been tested to reproduce values from a good binary of the master branch, but need more testing, expecially to test furhter possible speed optimizations. The inspiration for this came from the GROMACS table approach.