Reduced memory consumption, change to quad precision handling
Changed dynamic allocation of precomputed interaction list, now each node/core only allocates what it really needs.
This is the first step on the way of building real tables of only needed values, that can be automatically reduced during the simulation setup.
ALso changed it that quad precision handling is only used when desired by the user by passing the -DHAVEQUAD flag to the preprocessor