Bugfix and Charge group check
Bugfix for makefile, now also sets precision when building the debug
versions of the binaries.
Added check for charge group size against the currently used cut-offs.
Now Q will stop with an error if it detects charge groups that have a
larger maximum distance between the atoms than 75% of the lower value of
either the Protein-Water or Protein-Protein cut-off. This only affects
atoms not in the Q region and those that are not excluded. It will
prevent errors when using large molecules that have not been properly
cut into smaller charge groups.