Finished temperature control
Implementation of arbitrary temperature control groups is done now. The
control is done within the [TGROUPS] section, with the following
structure:

$firstatom $lastatom $groupname

Any number of groups can be defined, but the user should be aware that
the groups should not be too small or the scaling becomes meaningless.
Also, right now the program needs the exact atom numbers to be given,
and all atoms of the system need to be covered by the user. If no
section is found, the default groups are set up for solute and solvent
atoms automatically.