More temperature control changes
Added keyword
verbose_temp
in the MD section, deciding on the spamming of the Hot atoms warning and
the extra printing of temperatures if they excede the 2% cutoff in the
code. THis is useful for long runs, or if the minimization is not
expected to fail due to hot atoms.