LINCS now working
Fix for LINCS (and constraint handling in general). LINCS now works as intended,
used the wrong distance function before. General change to constraints that the sub arrays
of the constraint type are pointers to the actual arrays. It didn't feel right for me
to use implicit reference to those arrays, when reassigning the elements of the constraint array.