Important fix for organic solvents
Solvent forces on the boundary layer for organic solvents have so far
been to small by one over the number of heavy solvent atoms. In normal
water solvent, forces and energies are only calculated for the center,
heavy atom. In complex solvent, this has been done so far for the
geometric center of the molecule, but with the atom based forces scaled
down. Longer test simulation showed that this leads to solvent depletion
at the outer layers of the solvent shell. This update calculates the
forces on all heavy solvent atoms explicitly.
Also, some formatting changes.