Solvent generation change and ff upload
Changed solvent-solvent atom distance cutoff for deciding if new solvent
should be kept or not. Solvent-solute is still the same default of 2.4Å,
with the solvent-solvent value being half of it. This might change again
in the future with explicit set values. Also uploading the files for the
force field to be able to use solvents in oplsaa and fix to deprotonated
arginine and lysine torsions for oplsaa2015.