Option to use random numbers from system time
Added the option to use random number seeds from system time. This is
chosen if the user sets an initial temperature and either does not
specifiy a random number seed, or if the seed is set to less than zero.
This addresses issue #45, but will not be added to the Q5 master branch.