Q6 v6.0 Final Commit
This committ contains the final changes to Q to push it to version 6.0.
The changes are mostly related to the LICENCE statement, and the
addition of correct licencing information to all files. Also, files that
have been left over from previous versions and lingered on have been
removed. Finally, the manual has been updated for this version.